👌 IMP: Extract shared LR analysis code to analysis.rs
