+++ crates/princhess-train/src/lib.rs
@@ -3,6 +3,7 @@
 #![allow(clippy::cast_sign_loss, clippy::cast_possible_wrap)]
 #![allow(clippy::multiple_crate_versions)]
 
+pub mod analysis;
 pub mod args;
 pub mod data;
 pub mod eg_policy;
